+++ test_srcs/VectorTest.hpp
@@ -1,10 +1,15 @@
 #ifndef VectorTest_HPP
 #define VectorTest_HPP
 
+#include "UnitTester.hpp"
+#include "vector.hpp"
+#include <iostream>
 #include <list>
 #define TABLE_SIZE 0xF00
 
 typedef struct s_unit_tests t_unit_tests;
+
+namespace ft = std;
 
 namespace VectorTest {
 
@@ -53,6 +58,37 @@ void vector_operator_ge();
 void vector_std_swap();
 
 extern t_unit_tests func_test_table[TABLE_SIZE];
+
+void             _set_int_array(int* array, int size = 12, bool accend = false);
+ft::vector<int>  _set_vector(int size = 6, bool accend = false);
+ft::vector<char> _set_vector_char(size_t size = 6, bool accend = false);
+ft::vector<std::string>
+     _set_vector_string(size_t size = 6, bool random = false);
+void _set_compare_vectors(ft::vector<int>& ft_data, std::vector<int>& std_data,
+    int size = 6, bool accend = false);
+
+template <class T>
+void _compare_vectors(ft::vector<T>& ft_vec, std::vector<T>& std_vec)
+{
+	UnitTester::assert_(ft_vec.size() == std_vec.size());
+
+	size_t size = std_vec.size();
+	for (size_t i = 0; i < size; ++i) {
+		UnitTester::assert_(ft_vec[i] == std_vec[i]);
+	}
+}
+
+template <class T>
+void _debug(ft::vector<T>& ft_vec, std::string name = "")
+{
+	std::cerr << std::endl;
+	std::cerr << "[" << name << "] ";
+	size_t size = ft_vec.size();
+	for (size_t i = 0; i < size; ++i) {
+		std::cerr << i << ": " << ft_vec[i] << " ";
+	}
+	std::cerr << std::endl;
+}
 
 } // namespace VectorTest
